app/requirements.txt
fastapi
uvicorn
tensorflow
Pillow